fix some test cases
